// src/app/controles/control.interface.ts
export interface control{

    nombre:          string;
    centroDeTrabajo: string;
    mesesTrabajo:    Array<null | string>;
    lunes:           string[];
    martes:          string[];
    miercoles:       string[];
    jueves:          string[];
    viernes:         string[];
    sabado:          string[];
    domingo:         string[];


}